fix tick trade in beta
