+++ src/javascript/binary/websocket_pages/trade/get_ticks.js
@@ -1,7 +1,10 @@
 const TradingAnalysis      = require('./analysis');
 const Barriers             = require('./barriers');
 const TradingAnalysis_Beta = require('./beta/analysis');
+const Barriers_Beta        = require('./beta/barriers');
 const DigitInfo_Beta       = require('./beta/charts/digit_info');
+const Purchase_Beta        = require('./beta/purchase');
+const TickDisplay_Beta     = require('./beta/tick_trade');
 const updateWarmChart      = require('./common').updateWarmChart;
 const DigitInfo            = require('./charts/digit_info');
 const Purchase             = require('./purchase');
@@ -20,14 +23,19 @@ const GetTicks = (() => {
         }, {
             callback: (response) => {
                 const type = response.msg_type;
-                const is_digit = State.get('is_trading') && TradingAnalysis.getActiveTab() === 'tab_last_digit';
-                const is_digit_beta = State.get('is_beta_trading') && TradingAnalysis_Beta.getActiveTab() === 'tab_last_digit';
+                const is_digit = TradingAnalysis.getActiveTab() === 'tab_last_digit';
+                const is_digit_beta = TradingAnalysis_Beta.getActiveTab() === 'tab_last_digit';
                 if (type === 'tick') {
-                    processTick(response);
-                    if (is_digit) {
-                        DigitInfo.updateChart(response);
-                    } else if (is_digit_beta) {
-                        DigitInfo_Beta.updateChart(response);
+                    if (State.get('is_trading')) {
+                        processTick(response);
+                        if (is_digit) {
+                            DigitInfo.updateChart(response);
+                        }
+                    } else if (State.get('is_beta_trading')) {
+                        processTick_Beta(response);
+                        if (is_digit_beta) {
+                            DigitInfo_Beta.updateChart(response);
+                        }
                     }
                 } else if (type === 'history') {
                     processHistory(response);
@@ -69,10 +77,25 @@ const GetTicks = (() => {
         }
     };
 
+    const processTick_Beta = (tick) => {
+        const symbol = sessionStorage.getItem('underlying');
+        if (tick.echo_req.ticks === symbol || (tick.tick && tick.tick.symbol === symbol)) {
+            Tick.details(tick);
+            Tick.display();
+            TickDisplay_Beta.updateChart(tick);
+            Purchase_Beta.updateSpotList();
+            if (!Barriers_Beta.isBarrierUpdated()) {
+                Barriers_Beta.display();
+                Barriers_Beta.setBarrierUpdate(true);
+            }
+            updateWarmChart();
+        } else {
+            DigitInfo_Beta.updateChart(tick);
+        }
+    };
+
     return {
-        request       : request,
-        processTick   : processTick,
-        processHistory: processHistory,
+        request: request,
     };
 })();
